requirements.txt
# Azure SDK for Python - FinOps Project Dependencies
# Purpose: Azure cost management and resource inventory automation
# Install: pip install -r requirements.txt

# ==== Core (Required) ====
azure-identity>=1.15.0              # Authentication (DefaultAzureCredential, AzureCliCredential)
azure-mgmt-costmanagement>=4.0.0    # Cost extraction via Query API
azure-mgmt-resource>=23.0.0         # Resource inventory
pandas>=2.0.0                       # Data processing

# ==== Optional (Advanced Features) ====
# Uncomment if needed:
# azure-mgmt-consumption>=10.0.0      # Usage details, reservations
# azure-mgmt-subscription>=3.1.0      # Multi-subscription management
# azure-mgmt-advisor>=9.0.0           # Optimization recommendations
# openpyxl>=3.1.0                     # Excel export
